+++ CHANGELOG.md
@@ -11,10 +11,15 @@ unreleased
 
 ## Changed
 
+- wamp_session::provide() accept callback that is invoked on success/failure of
+  registration.
+
 - Using msgpack-c 2.1.2.  In this version the header-only msgpack decoder has
   been fixed, which is used by wampcc.
 
 ## Fixed
+
+- wamp_session not handling failed registration
 
 - compile errors on Xcode 7.3.1 + 10.7 SDK (issue #2, petten)